### 2019-01-21 #1192 @whedon
[PRE REVIEW]: CluSim: a Python package for the comparison of clusterings and dendrograms
**Submitting author:** @ajgates42 (<a href="http://orcid.org/0000-0003-0099-7480">Alexander Gates</a>)
**Repository:** <a href="https://github.com/Hoosier-Clusters/clusim" target ="_blank">https://github.com/Hoosier-Clusters/clusim</a>
**Version:** v0.3.1
