### 2019-01-21 #1192 @arfon
@ajgates42  - looks like there's a whitespace issue in your `paper.md`:

```
affiliations:
 - name: Department of Physics, Northeastern University, Boston, 02115, USA
   index: 1
 - name: Department of Informatics, Indiana University, Bloomington, 47408, USA
   index: 2
- name: Program in Cognitive Science, Indiana University, Bloomington, 47408, USA
   index: 3
```

Could you update it to make the indentation of each of the affiliations the same?
