### 2019-01-22 #1192 @arfon
Not quite @ajgates42 - this should fix your paper: https://github.com/Hoosier-Clusters/clusim/pull/24
