### 2019-01-22 #1192 @arfon
BTW, once you've made the changes to your `paper.md` file in `master`, you can ask @whedon to recompile the paper here with `@whedon generate pdf`
