### 2019-01-22 806643d
update figures
~ increased png resultion
~ grouped caption with image
