### 2019-01-22 #1192 @ajgates42
On Tue, Jan 22, 2019 at 11:50 AM Arfon Smith <notifications@github.com>
wrote:

> BTW, once you've made the changes to your paper.md file in master, you
> can ask @whedon <https://github.com/whedon> to recompile the paper here
> with @whedon generate pdf
>
