### 2019-01-22 #1192 @ajgates42
perfect, thank you!

On Tue, Jan 22, 2019 at 11:49 AM Arfon Smith <notifications@github.com>
wrote:

> @ajgates42 <https://github.com/ajgates42> - you need to put the caption
> text actually in the image tag e.g.:
>
> ![**Evaluating clustering comparisons w.r.t. random models.** A comparison using the Rand Index between the classification of cancer types and clustering labels derived using Hierarchical Clustering on gene expression data ($0.5$, red). above, Pairwise comparisons between samples from the Permutation model (blue, see [@Hubert1985adjrand]) with mean $0.44$ (black). below, Pairwise comparisons between samples from the one-sided model with a Fixed Number of Clusters (blue, see [@Gates2017impact]) with mean $0.59$ (black). The Permutation model suggests Hierarchical Clustering is more similar to the ground truth than a random clustering, while the one-sized fixed number of clusterings model, the more appropriate model for this scenario, reveals that the result is less similar than random clusterings.](paperfigures/CluSimFig1.png)
>
