### 2019-01-22 #1192 @mrosvall
I am on paternity leave and will not be able to provide a review.

Best regards,

Martin

> On 22 Jan 2019, at 18:26, Alexander Gates <notifications@github.com> wrote:
>
> Hi Viviane,
>
> Cool! Here are a few potential reviewers:
>
> Prof. Martin Rosvall <http://www.tp.umu.se/~rosvall/> @mrosvall
> Dr. Vinh Nguyen <https://vinhnguyenx.net> @vinhngx
> Prof. Marcilio de Souto
> <http://www.univ-orleans.fr/lifo/Members/Marcilio.DeSouto/>
> Dr. Thomas Aynaud <https://perso.crans.org/aynaud/> @taynaud
>
>
> Cheers,
> Alex
>
>
>
> On Tue, Jan 22, 2019 at 12:06 PM Viviane Pons <notifications@github.com>
> wrote:
>
> > Perfect :) I am looking for reviewers, if you have any suggestion, please
> > let me know.
> >
> > —
> > You are receiving this because you were mentioned.
> > Reply to this email directly, view it on GitHub
> > <https://github.com/openjournals/joss-reviews/issues/1192#issuecomment-456479995>,
> > or mute the thread
> > <https://github.com/notifications/unsubscribe-auth/AZk97XGvltpCxF0FlCvyH2o4CwaGMSkwks5vF0UigaJpZM4aKs7X>
> > .
> >
>
>
> --
> Alexander Gates
> Postdoctoral Research Scholar
> Center for Complex Network Research
> Northeastern University
> http://alexandergates.net
