### 2019-01-23 #1192 @ajgates42
Cool, thanks Viviane.  I'm impressed with how quickly this process is
progressing.

On Wed, Jan 23, 2019 at 9:25 AM Viviane Pons <notifications@github.com>
wrote:

> We have one reviewer, I would feel more comfortable with 2. I have just
> sent an email to Marcillo de Souto to see if he is interested. Once I have
> two confirmed reviewers, I will create the review issue.
>
