### 2019-01-24 #1192 @adavidzh
In order to assess whether I could perform this review, I went through:
1. the [docs](https://hoosier-clusters.github.io/clusim/html/clusim.html),
1. the [`2018 bioRxiv`](https://doi.org/10.1101/410084), and
1. the [`arXiv:1706.06136`](https://arxiv.org/abs/1706.06136).

<summary>(remainder of post collapsed to not bias other reviewers)</summary>
I could only understand what was being done when I got to the `arXiv` paper.

As far as I can tell, the `JOSS` submission is nearly identical to the `bioRxiv` paper.

In the spirit of speeding things along, I would like to note that I see an issue with the examples:
- the docs have a dry and contrived example (that I do not understand), and
- the two examples provided in the `bioRxiv` paper are impossible to decode for a broader audience.

At variance with that, the `arXiv` paper is a pleasure to read.
I would go as far as to say that `CluSim` comes out of the benchmarking done in that paper for the element-centric framework, comparing it with other measures of cluster similarity.

I think that in order to make the `CluSim` `JOSS` article useful for (potentially interested) users, it should use as examples the results for the four scenarios and multiple metrics that are shown in `Figs. 2a-c` and `S1-3` of the `arXiv` paper, instead of the ones presently used.

I understand that we are still on the pre-review stage and authors should wait to take action on this.
