### 2019-02-06 #1192 @ajgates42
OK, thanks Viviane.  I have thought about Andre's comments, but was waiting
for the official review process to start before making any changes.

Cheers,
Alex

On Wed, Feb 6, 2019 at 12:52 PM Viviane Pons <notifications@github.com>
wrote:

> Just to let you know that I'm still in the process of finding a second
> reviewer.
>
