### 2019-02-19 #1264 @adavidzh
I have a few fundamental comments [from the pre-review](https://github.com/openjournals/joss-reviews/issues/1192#issuecomment-457099152) which I would hope the authors can comment on.
