### 2019-03-13 530e7cb
improved examples of the package
~ modified the paper to include behavior examples from "On comparing clusterings"

~ include a new example notebook that creates these figures
