### 2019-03-13 #1264 @ajgates42
Thanks @adavidzh for the close reading of our work!  I understand the lack of details in our first write-up made it difficult to decode the examples, we were trying to illustrate the breadth of the package which staying within the constraints of the word limits.

To improve the readability and ease of use, we have followed your suggestion and exchanged one of those examples with a simplified version of the Arxiv behavior examples.  We chose to retain the random model example because it illustrates another core functionality of the package, but we tried to simplify the explanation in the writeup.

I also included a new [example jupyter notebook](https://github.com/Hoosier-Clusters/clusim/blob/joss-reviews/examples/CluSim%20JOSS%20Writeup%20Examples.ipynb) that recreates the examples discussed in the text.

All updates can be found in the [joss-reviews branch](https://github.com/Hoosier-Clusters/clusim/tree/joss-reviews).
