### 2019-03-14 #1264 @adavidzh
Quick question for @VivianePons: the review checklist is addressed to @pajaskowiak. Does this mean that only they should check items or can I also do it?
