### 2019-03-15 #1264 @pajaskowiak
Just read the manuscript. Excelent work, easy to follow and understand. My only minor point is w.r.t. the fact that the authors mention that all measures are bounded to the [0,1] interval. This is not the case for Adjusted Rand Index (ARI), which can be negative. Other than that, I have no concerns at all.
