### 2019-03-17 #1264 @VivianePons
The command added you in the list but didn't create the list of checkboxes, I just did manually. Please, just check that you can check them all ;)

Thanks and sorry for the confusion!
