### 2019-03-17 #1264 @adavidzh
@VivianePons: all good, thanks.
@ajgates42: while going through the review checklist:
1. Can you point me to the contribution guidelines? IMO, this can be as simple as adding a section to `README.md`.
1. I can't quite find how the `0.3.1` version is tagged: I see that version in [PyPI](https://pypi.org/project/clusim/) but there is no branch nor tag that I can find on [github](https://github.com/Hoosier-Clusters/clusim/branches/all).
