### 2019-03-18 21020be
update code of conduct and contribution
version 0.3.2
