### 2019-03-18 #1264 @ajgates42
@adavidzh, you can now find a contribution guidelines and code of conduct in the [README](https://github.com/Hoosier-Clusters/clusim/blob/master/README.md).  I've also tagged a new release v0.3.2 and updated the version in the __init__.py.  The PyPI version will be updated shortly.
Thanks!
