### 2019-03-19 #1264 @adavidzh
> I'm sorry @adavidzh, I'm afraid I can't do that. That's something only editors are allowed to do.

Hi @VivianePons, is it correct to say that the version number should change, given the [additions to the repo](https://github.com/Hoosier-Clusters/clusim/tree/0.3.2)?
