### 2019-03-19 #1264 @VivianePons
Yes, it seems to make sense. I just updated it.

Do we have a green light from both reviewers now to accept the paper?
