### 2019-03-19 #1264 @VivianePons
Perfect!

@ajgates42 could you create an archive of your repo on Zenodo or figshare and generate a DOI? Thanks
