### 2019-03-19 #1264 @ajgates42
Great!  Thank you all for the valuable feedback!

DOI: https://doi.org/10.5281/zenodo.2598682
Record: https://zenodo.org/record/2598682

On Tue, Mar 19, 2019 at 1:13 PM Viviane Pons <notifications@github.com>
wrote:

> Perfect!
>
> @ajgates42 <https://github.com/ajgates42> could you create an archive of
> your repo on Zenodo or figshare and generate a DOI? Thanks
>
