### 2019-03-19 #1264 @VivianePons
@ajgates42 it seems like some references are missing their dois. For example, I see: "The impact of random models on clustering similarity" has doi https://doi.org/10.1101/196840 which is not in the bib file. Can you check? whedon has put the list of all references without dois.

Thanks
