### 2019-03-19 #1264 @arfon
@ajgates42 - you might also want to remove entries from your bibtex that you're not actually using in this paper?
