+++ paper.bib
@@ -1,81 +1,3 @@
-@article{Jaccard1912flora,
-  author = {Jaccard, Paul},
-  title = {The distribution of the flora in the alpine zone},
-  journal = {New Phytologist},
-  uuid = {3C0D4AEF-2DC9-46DE-8F86-82B2D4EAF5DD},
-  volume = {11},
-  number = {2},
-  pages = {37-50},
-  publisher = {Wiley Online Library},
-  year = {1912}
-}
-
-
-@article{Marczewski1958setdistance,
-  title = {On a certain distance of sets and the corresponding distance of functions},
-  volume = {6},
-  number = {1},
-  journal = {Colloquium Mathematicae},
-  author = {Marczewski, Edward and Steinhaus, Hugo},
-  year = {1958},
-  pages = {319--327}
-}
-
-@article{Levandowsky1971jaccardmetric,
-  author = {Levandowsky, Michael  and Winter, David},
-  title = {Distance between sets},
-  journal = {Nature},
-  uuid = {C02BFEA1-917B-4A85-B391-039C9A7D478D},
-  volume = {234},
-  number = {5323},
-  pages = {34-35},
-  publisher = {Nature Publishing Group},
-  year = {1971}
-}
-
-
-@article{Rand1971randindex,
-    author = {Rand, William M},
-    title = {{Objective Criteria for the Evaluation of Clustering Methods}},
-    journal = {Journal of the American Statistical Association},
-    year = {1971},
-    volume = {66},
-    number = {336},
-    pages = {846}
-}
-
-@article{Stone1974cross,
-  title={Cross-validatory choice and assessment of statistical predictions},
-  author={Stone, Mervyn},
-  journal={Journal of the Royal Statistical Society. Series B (Methodological)},
-  pages={111--147},
-  year={1974},
-  publisher={JSTOR}
-}
-
-@article{Waterman1978simdendro,
-  author = {Waterman, Michael S.  and Smith, Temple F.},
-  title = {On the similarity of dendrograms},
-  journal = {Journal of Theoretical Biology},
-  uuid = {8973AF30-D508-423E-98CE-F2ECC6CD729F},
-  volume = {73},
-  number = {4},
-  pages = {789-800},
-  publisher = {Elsevier},
-  year = {1978}
-}
-
-@article{Fowlkes1983hierarchicalcompare,
-  author = {Fowlkes, Edward B.  and Mallows, Colin L.},
-  title = {A method for comparing two hierarchical clusterings},
-  journal = {Journal of the American Statistical Association},
-  uuid = {CB4E29B8-FECB-4E41-9FEE-98988D9E2273},
-  volume = {78},
-  number = {383},
-  pages = {553-569},
-  publisher = {Taylor \& Francis Group},
-  year = {1983}
-}
 
 @article{Hubert1985adjrand,
 author = {Hubert, Lawrence and Arabie, Phipps},
@@ -85,51 +7,8 @@ year = {1985},
 volume = {2},
 number = {1},
 pages = {193--218},
-month = dec
-}
-
-@article{Collins1988omega,
-  author = {Collins, Linda M.  and Dent, Clyde W.},
-  title = {Omega: A general formulation of the rand index of cluster recovery suitable for non-disjoint solutions},
-  journal = {Multivariate Behavioral Research},
-  uuid = {39B3538A-4CC2-46E5-B054-3DCE5890A2E0},
-  volume = {23},
-  number = {2},
-  pages = {231-242},
-  publisher = {Taylor \& Francis},
-  year = {1988}
-}
-
-@article{White1994nmibias,
-  author = {White, Allan P.  and Liu, Wei Zhong},
[…]
