+++ paper.bib
@@ -53,7 +53,8 @@ doi = {10.1007/BF01908075}
 	journal={Journal of Machine Learning Research},
 	author={Gates, Alexander J. and Ahn, Yong-Yeol}, 
 	year={2017}, 
-	pages={1-28}
+	pages={1-28},
+  doi = {10.1101/196840}
  }
 
 @misc{Gates2018element,
