### 2019-03-19 #1264 @ajgates42
@VivianePons <https://github.com/VivianePons>: JMLR actually doesn't
actually give a DOI yet, but I guess we can use the BioRxiv version instead.

On Tue, Mar 19, 2019 at 5:04 PM Arfon Smith <notifications@github.com>
wrote:

> @ajgates42 <https://github.com/ajgates42> - you might also want to remove
> entries from your bibtex that you're not actually using in this paper?
>
