### 2019-03-19 #1264 @VivianePons
We're still missing a doi for "Data clustering: a review" (not listed by whedon for some reason...). Other missing ones are from "misc" paper, arxiv preprint, so I guess this is ok.

Also the date on the scipy citation prints weirdly, @arfon do you know if it's the correct way to do it?
