+++ paper.bib
@@ -20,7 +20,8 @@ doi = {10.1007/BF01908075}
   number = {3},
   pages = {264-323},
   publisher = {Acm},
-  year = {1999}
+  year = {1999},
+  doi = {10.1145/331499.331504}
 }
 
 @article{deSouto2008geneclustering,
@@ -46,13 +47,13 @@ doi = {10.1007/BF01908075}
   doi = {10.1145/1553374.1553511}
 }
 
-@article{Gates2017impact, 
-	title={The Impact of Random Models on Clustering Similarity}, 
-	volume={18}, 
-	number={87}, 
+@article{Gates2017impact,
+	title={The Impact of Random Models on Clustering Similarity},
+	volume={18},
+	number={87},
 	journal={Journal of Machine Learning Research},
-	author={Gates, Alexander J. and Ahn, Yong-Yeol}, 
-	year={2017}, 
+	author={Gates, Alexander J. and Ahn, Yong-Yeol},
+	year={2017},
 	pages={1-28},
   doi = {10.1101/196840}
  }
