### 2019-03-20 #1264 @arfon
> Also the date on the scipy citation prints weirdly, @arfon do you know if it's the correct way to do it?

Not sure, I think it's probably because of the bibtex entry not actually having a date range https://github.com/Hoosier-Clusters/clusim/blob/master/paper.bib#L85
