### 2019-03-20 dc7cb61
removed scipy citation
