+++ paper.bib
@@ -78,14 +78,6 @@ doi = {10.1007/BF01908075}
  year={2011}
 }
 
-
-@Misc{scipy,
-  author =    {Eric Jones and Travis Oliphant and Pearu Peterson and others},
-  title =     {{SciPy}: Open source scientific tools for {Python}},
-  year =      {2001--},
-  url = "http://www.scipy.org/"
-}
-
 @article{Sukumaran2010dendropy,
 author = {Sukumaran, J. and Holder, Mark T.},
 title = {DendroPy: A Python library for phylogenetic computing},
