### 2019-03-20 #1264 @ajgates42
Thats the standard citation from the scipy website
<https://www.scipy.org/citing.html>.  But the reference isn't essential for
the work, so lets, just drop it from the paper.  I've updated the bib
accordingly.

Thanks!

On Wed, Mar 20, 2019 at 6:10 AM Arfon Smith <notifications@github.com>
wrote:

> Also the date on the scipy citation prints weirdly, @arfon
> <https://github.com/arfon> do you know if it's the correct way to do it?
>
> Not sure, I think it's probably because of the bibtex entry not actually
> having a date range
> https://github.com/Hoosier-Clusters/clusim/blob/master/paper.bib#L85
>
