### 2019-03-20 #1264 @VivianePons
Perfect, could you update the Zenodo metadata so that the title matches the paper name? (At the moment, the Zenodo title is only "CluSim" whereas the title paper is: "CluSim: a Python package for the comparison of clusterings and dendrograms")

Also, you might wanna udpate the paper.md and paper.bib on the the Zenodo archive if that's possible (sorry, I should probably have asked you to create the archive AFTER checking the references)
