### 2019-03-21 #1264 @ajgates42
OK.  The Zenodo metadata and files have been brought up to date.  The new
DOI is 10.5281/zenodo.2601868

On Wed, Mar 20, 2019 at 5:50 PM Viviane Pons <notifications@github.com>
wrote:

> Perfect, could you update the Zenodo metadata so that the title matches
> the paper name? (At the moment, the Zenodo title is only "CluSim" whereas
> the title paper is: "CluSim: a Python package for the comparison of
> clusterings and dendrograms")
>
> Also, you might wanna udpate the paper.md and paper.bib on the the Zenodo
> archive if that's possible (sorry, I should probably have asked you to
> create the archive AFTER checking the references)
>
