### 2019-03-21 #1264 @ajgates42
Awesome!  Thank you everyone for the insightful feedback and great
submission experience.

On Thu, Mar 21, 2019 at 1:21 PM Arfon Smith <notifications@github.com>
wrote:

> @pajaskowiak <https://github.com/pajaskowiak>, @adavidzh
> <https://github.com/adavidzh> - many thanks for your reviews here and to
> @VivianePons <https://github.com/VivianePons> for editing this submission
> ✨
>
> @ajgates42 <https://github.com/ajgates42> - your paper is now accepted
> into JOSS ⚡️🚀💥
>
